+++ internal/bundler/bundler_ts_test.go
@@ -607,11 +607,13 @@ func TestTypeScriptDecorators(t *testing.T) {
 					@x @y mUndef
 					@x @y mDef = 1
 					@x @y method(@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
+					@x @y declare mDecl
 					constructor(@x0 @y0 arg0, @x1 @y1 arg1) {}
 
 					@x @y static sUndef
 					@x @y static sDef = new Foo
 					@x @y static sMethod(@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
+					@x @y static declare mDecl
 				}
 			`,
 			"/all_computed.ts": `
@@ -621,6 +623,7 @@ func TestTypeScriptDecorators(t *testing.T) {
 					@x @y [mUndef()]
 					@x @y [mDef()] = 1
 					@x @y [method()](@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
+					@x @y declare [mDecl()]
 
 					// Side effect order must be preserved even for fields without decorators
 					[xUndef()]
@@ -631,6 +634,7 @@ func TestTypeScriptDecorators(t *testing.T) {
 					@x @y static [sUndef()]
 					@x @y static [sDef()] = new Foo
 					@x @y static [sMethod()](@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
+					@x @y static declare [mDecl()]
 				}
 			`,
 			"/a.ts": `